Move telemetry files into settings package, delete telemetry package
The internal/cmd/telemetry package is removed entirely. The three files
are recreated as package settings under internal/cmd/settings/, and
settings.go now calls newTelemetryCmd() directly with no cross-package
import.
