+++ internal/cmd/settings/telemetry_disable.go
@@ -20,7 +20,7 @@
 //
 // SPDX-License-Identifier: MIT
 
-package telemetry
+package settings
 
 import (
 	"context"
@@ -33,38 +33,40 @@ import (
 	"github.com/CircleCI-Public/circleci-cli/internal/iostream"
 )
 
-func newEnableCmd() *cobra.Command {
+func newTelemetryDisableCmd() *cobra.Command {
 	cmd := &cobra.Command{
-		Use:   "enable",
-		Short: "Opt in to telemetry",
+		Use:   "disable",
+		Short: "Opt out of telemetry",
 		Long: heredoc.Doc(`
-			Enable anonymous usage telemetry for the CircleCI CLI.
+			Disable anonymous usage telemetry for the CircleCI CLI.
 
-			This preference is saved to the CLI config file. Environment variables
-			(CIRCLECI_NO_TELEMETRY, NO_ANALYTICS, DO_NOT_TRACK, CI) always take
-			precedence and will disable telemetry even when this setting is enabled.
+			This preference is saved to the CLI config file. You can re-enable
+			telemetry at any time with 'circleci settings telemetry enable'.
+
+			Alternatively, set CIRCLECI_NO_TELEMETRY, NO_ANALYTICS, or DO_NOT_TRACK
+			in your environment to disable telemetry without modifying the config file.
 		`),
 		Example: heredoc.Doc(`
-			# Opt in to telemetry
-			$ circleci settings telemetry enable
+			# Opt out of telemetry
+			$ circleci settings telemetry disable
 
-			# Opt in using a custom config file path
-			$ circleci --config ~/.config/circleci/config.yml settings telemetry enable
+			# Opt out using a custom config file path
+			$ circleci --config ~/.config/circleci/config.yml settings telemetry disable
 
-			# Verify the stored preference after enabling
-			$ circleci settings list --json
+			# Disable telemetry for a single session without changing the config
+			$ CIRCLECI_NO_TELEMETRY=1 circleci pipeline list
 		`),
 		Args: cobra.NoArgs,
 		RunE: func(cmd *cobra.Command, args []string) error {
 			ctx := iostream.FromCmd(cmd.Context(), cmd)
 			configPath, _ := cmd.Flags().GetString("config")
-			return runEnable(ctx, configPath)
+			return runTelemetryDisable(ctx, configPath)
 		},
 	}
 	return cmd
 }
 
-func runEnable(ctx context.Context, configPath string) error {
+func runTelemetryDisable(ctx context.Context, configPath string) error {
 	resolvedPath := configPath
 	if resolvedPath == "" {
 		var err error
@@ -75,16 +77,12 @@ func runEnable(ctx context.Context, configPath string) error {
 		}
 	}
 
-	if err := config.SetTelemetryEnabled(ctx, true, resolvedPath); err != nil {
+	if err := config.SetTelemetryEnabled(ctx, false, resolvedPath); err != nil {
 		return clierrors.New("telemetry.save_failed", "Failed to save telemetry setting", err.Error()).
 			WithExitCode(clierrors.ExitGeneralError)
 	}
 
-	iostream.ErrPrintf(ctx, "%s Telemetry enabled. Saved to %s\n", iostream.SymbolOK(ctx), resolvedPath)
-
-	for _, env := range config.ActiveTelemetryOverrides() {
-		iostream.ErrPrintf(ctx, "Note: %s is set — telemetry remains disabled for this session.\n", env)
-	}
+	iostream.ErrPrintf(ctx, "%s Telemetry disabled. Saved to %s\n", iostream.SymbolOK(ctx), resolvedPath)
 
 	return nil
 }
